refactor(theme): align provider path with plan
Rename the appearance provider to theme-provider and update imports so the implementation matches the planned file structure without changing behavior.

Made-with: Cursor
